requirements.txt
# HistoLab - Histopathology Image Classification Dependencies

# Core ML/AI
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
trl>=0.7.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# UI
gradio>=4.0.0

# Configuration
PyYAML>=6.0

# Scientific Computing & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0

# Dataset & Model Hub
datasets>=2.14.0
huggingface_hub>=0.19.0

# Experiment Tracking
wandb>=0.16.0

# Evaluation Metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
